@@ -1,7 +1,7 @@
 package com.openci.core;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import com.openci.R;
 import com.openci.ui.fragments.RepositoriesFragment;
