@@ -7,6 +7,7 @@ import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Header;
 import retrofit2.http.Headers;
+import retrofit2.http.Query;
 
 /**
  * Created by Vicky on 08-01-2018.
@@ -18,5 +19,5 @@ public interface IRepos {
             "User-Agent: MyClient/1.0.0"
     })
     @GET("repos")
-    Call<ReposResponse> getUser(@Header("Authorization") String authorization);
+    Call<ReposResponse> getRepos(@Header("Authorization") String authorization, @Query("include") String include);
 }
